Add queries

+++ app/src/main/java/com/android/squadster/ui/usersquad/recyclerview/MemberViewHolder.kt
@@ -43,7 +43,7 @@ class MemberViewHolder(
             }
             itemView.iv_delete.setOnClickListener {
                 member.id.let {
-                    handler.deleteMember(member.id.toInt())
+                    handler.deleteMember(member.id)
                 }
             }
         } else {
